Moving UserRepository related code into their own file
