+++ src/repositories/interfaces/UserRepository.ts
@@ -1,0 +1,9 @@
+export interface UserRepository {
+  findById(id: string): Promise<UserEntity | null>;
+}
+
+export type UserEntity = {
+  id: string;
+  name: string;
+  profileImageUrl: string;
+};
